kadai

+++ app/Http/Controllers/TasksController.php
@@ -5,7 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Task;    // 追加
-
+use App\User; // 追加
+use Illuminate\Support\Facades\Auth;
 class TasksController extends Controller
 {
     /**
@@ -13,12 +14,23 @@ class TasksController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
     public function index()
     {
-        $task = Task::all();
-        return view('tasks.index', [
-            'task' => $task,
-        ]);
+        // 認証済みユーザを取得
+        $user = Auth::user();
+        if($user){
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $query = $user->task()->orderBy('created_at', 'desc')->paginate(10);
+
+            $task = [
+                'user' => $user,
+                'task' => $query,
+            ];
+            return view('tasks.index', $task);
+        }else{
+            return view('welcome');
+        }
     }
 
     /**
@@ -42,20 +54,19 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        
         // バリデーション
         $request->validate([
             'status' => 'required|max:10',
         ]);
-        
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
 
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->task()->create([
+            'content' => $request->content,
+            'status' => $request->status
+        ]);
         return redirect('/');
     }
-    
+
     /**
      * Display the specified resource.
      *
@@ -64,10 +75,31 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::findOrFail($id);
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        // idの値でユーザを検索して取得
+        // 認証済みユーザを取得
+        $user = Auth::user();
+		//$user = User::findOrFail($id);
+        $user->loadRelationshipCounts();
+
+        // ユーザの投稿一覧を作成日時の降順で取得
+        $query = $user->task()->where('id', $id)->orderBy('created_at', 'desc')->paginate(10);
+        $post=$query['0'];//投稿データ
+
+        $task = [
+            'user' => $user,
+            'task' => $post,
+        ];
+        //$task2 = \App\task::findOrFail($id);
+	if(isset($post)){//投稿があるかどうか
+		if(\Auth::id() === $query['0']->user_id){
+        		return view('tasks.show', $task);
+		}else{
+	        	return redirect('/');
+		}
+	}else{
+		return redirect('/')->with('message', 'タスクは存在しません');
+	}
+
     }
     /**
      * Show the form for editing the specified resource.
@@ -77,10 +109,24 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::findOrFail($id);
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+    	$user = Auth::user();
+    	$query = $user->task()->where('id', $id)->orderBy('created_at', 'desc')->paginate(10);
+    	$post=$query['0'];//投稿データ
+
+    	$task = [
+    			'user' => $user,
+    			'task' => $post,
+    	];
+        //$task = Task::findOrFail($id);
+        if(isset($post)){//投稿があるかどうか
+			if(\Auth::id() === $query['0']->user_id){
+	  	      return view('tasks.edit', $task);
+			}else{
+	 	       return redirect('/');
+			}
+       	}else{
+       		return redirect('/')->with('message', 'タスクは存在しません');
+        }
     }
 
     /**
@@ -95,11 +141,11 @@ class TasksController extends Controller
         // バリデーション
         $request->validate([
             'status' => 'required|max:10',
-        ]);        
-        
+        ]);
+
         $task = Task::findOrFail($id);
         $task->content = $request->content;
-        $task->status = $request->status;        
+        $task->status = $request->status;
         $task->save();
         return redirect('/');
     }
@@ -112,8 +158,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\task::findOrFail($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         return redirect('/');
     }
 }
